Add flask api and organized core logic better
Adds a webapp on top of our core logic in order to have an API to
interact with the hangman game. The api description can be seen in
the README.md. Also separates some files and removes the guesses.py
file in order to improve clarity. This commit finishes all the work
on the backend.